Implement RayTracing including antialiasing
The code is nasty of cource, but that is besides the point.
For the time being I am only interested in playing around with the concepts
following the tutorial and getting a feel for the beauty of raytracing.
If and when I have enough confidence to start making the desired product
I will start doing so.