Add BVH Volumes support
Now this structure is rather suboptimal as the comparison function
itself will construct a bounding box for each comparison and then
finally again another one will be built in the end.
Non the less we will stick with it as it is how it is in the tutorial
and it is unclear yet where the code will go.