Add lina - linear algebra  library
As a start it is as simple as it can be. It only implements the
smallest set of operations we need on the strictest types.
Initially it only needs to support 3 long inputs, and only
return 3 long outputs. That is it, no need even for templates,
as we simply use doubles.
As simple as it gets. Easy to test, easy to understand.