Separate out the Camera class
In its current form the Camera class is barely useful, but it
already demonstrates one key observation. It has nothing to do
with how the image itself is generated. It has nothing to do
with threading or color generation etc. It only needs to know
where it is, where it is looking at, and the screen resolution.
Then it only needs to generate the matrix of rays and give those back
when requested. That is all it needs to know.

Later this can be extended by anti aliasing, where it would generate
more ray. Another feature I really want is to be able to freely move
it, rotate it etc. For that it would need to receive a lot more
arguments, but for now it is okay as it is.