Add Lambertian material and Material
Material represents the material properties of a given object.
It is a separate entity and does not need to know about the shape of
the object. Geometry decides where an incoming ray hits, where the normal
is, etc. Material describes how that incoming ray would then bounce off
in which directions, with which attenuation etc.