Move collision point adjustment out of geometry
While implementing the Dialectric material I have realized
that it should be up to the material to decide how the collision
point should be adjusted. This is because a valid such point
will depend on whether or not we have hit the objects boundary
from the outside or the inside, then where they ray should go next.