Fix calculation errors on Lambertian, Metal materials
In my haste the Metal material code was rather broken. But apart
from that, the code have been adjusted such that if our randomly
perturbed calculations would produce an invalid result, we simply
reproduce the random event. The reasoning is that if we loose
a reflection ray just because of the random event producing an invalid
result, we have cut short the path tracing process, which introduces
more error into the computations, requiring more overal samples and
longer times to converge on the result.