Add a simple plane geometry
This plane can only have one orientation, aligned along the x, z
axis and can only have it's normal pointed up.
May seem restrictive at first, but the goal is simplicity. With
the camera we can already see how very unwieldy it is having to define an
up vector. It would be much easier if rather we had a very restrictive way
of constructing the plane and then rotating/translating it into its final
position. I plan to have rotations and translations anyways.