Add design documentation
Just recently occured to me that I should add a design document
describing what is being made here, what is the end goal and how it should be
done. Didn't go into too much detail, for once, because I don't know about the
existence of those details, for second, because it is not necessary.

The aim is to define the goals cleanly and clearly. As a start it doesn't have
to be more complicated than that.