Add rotation support for geometry
Modified the plane implementation a bit, so that the transformations
may be uniformly applied to both the center and the direction vectors
of. The trick is to store both the center and the "side" vectors of
the plane in the global space.
Previously only the center was in global space, while the sides were
locally defined from the center.
By keeping them uniform it, doesn't matter what type of transformation comes
in, it should be applied to all vectors the same.

Noticed one oversight, while trying to add the transisiont transformation.
In matrix form a transition needs a 4x4 matrix, so we have to work with 4 element
vectors. The refactor shouldn't take much effort, as the final element for such
4 long vectors is always one.