Extend transformation matrices to 4x4
To be able to implement translation in matrix form we need
4x4 matrices. This adissional dimension allows us to distinguish
between directions and points. Translation only applies to points
, while every other operation (scaling, rotation, shearing) applies
to both point and direction vectors. Now we can finally combine such
transformations together and start descibing what we need to.
Ahh, this fills me with joy. Finally!

Unfortunatelly, now we have to transition into 4D vectors, to have
everything consistent. This shouldn't be an issue, just a bit of a hassle
to adjust everything. Oh well, either we know, or we learn.