Minor cleanups in design document
For a moment there I was confused about which handedness to use in the
application and wanted to switch to a left handed orientation.
Then checked what does Vulkan use, and that uses the right hand orientation.
The only difference now between the coordinate system here and in Vulkan is the 180
degree rotation around the X axis. Which is simple to adjust for.