Move vector extension/contraction into transform.h
At this point it isn't entirely clear to me that we should be working
with 3 or 4 dimensional vectors. To be able to handle linear transformations
uniformly we need 4D vectors. But the 4th dimension is only to distinguish
between direction and position vectors. Otherwise it has no real meaning.

Furthermore, cross product is not defined on 4D vectors, at least what exists
wouldn't be what we would like to generate, to my current understanding.
For the time being, when needed we will simply extend and cut vectors
so that they fit the required format at any given time.
Maybe in the future some cleaner way will represent itself.