Add an option to easily build a plane in cardinal orientations
Additionally added the Cornell box scene. There is one thing
I have realized. Why most meshes are made out of triangles.
Well, at least it is my theory based on what I have understood so far.
If you have a mesh, any mesh, that consists of triangles, which are defined
by offset vectors around a central point, then you can easily rotate, scale
it in any direction. It will be just multiplying the vectors by the transformation
matrix.
Realized this after it occurred to me that my original plan of defining a cuboid
glass using these planes is bullcrap. The reason is that although I have
scalling implemented on them, and I could make a cuboid using a center point
and six limited planes, it wouldn't be just as simple as multiplying all
vectors by the transformation matrix. The reason is that scalling a side
appropliately has to take into account how far we are from the center.
Oh well, live and learn.