Refactor sample ray generation
With the original way, when we pre-generated all sample rays 10 GiB
of memory was used for 300 samples at 1024x768 resolution.
A little too much, and it clearly won't scale for higher details.
After the refactor, memory use is constant and with the same setup
the whole process runs about 30 seconds faster.
Oh well, stupid original idea, but fixed rather easily.