Add hardening tooling to the project
Added clang-tidy, but disabled a few annoying checks like the magic
numbers, short names and recursion. These were a little too strict
for a path tracer.
Unfortunatelly clang-tidy is a little wonky, because we are on
c++2b experimental and for example std::expected is frowned upon.
Oh well, we fixed all the issues we could.

Added hedron which can generate compile commands from bazel projects.