Add simple script to call clang-tidy on the sources
Marginally useful for the time being. I need to setup a bazel toolchain
using clang, setting the appropriate compile flags and then it may be
used in CL. Until then it is problematic, because by default gcc is used
for which I hacked the flag in, but tidy will try to use clang.
And it won't find the right version either.
Regardless it caught a number of bugs already so it is still useful,
but not yet up to its full potential.