Add PDF support for materials
Now the materials that can support the usage of PDFs. The current
implementation doesn't change anything about the rendering itself,
because what we have is equivalent to the prior version.
But it enables us to extend it further in the next change by adding
PDF support for sampling the geometries as well, which combined
with a mixture PDF, will significantly improve the quality and
speed of our renders.