Added mixed pdf handling, but something is wrong
There are multiple issues with this solution. The biggest one
is that it requires additional code for calculating the sampling
pdf for that object. Combining then multiple PDF received this
way doesn't take into account if their sampling direction overlaps.
If it does than our math gets scewed, because currently our math
behaves as if these pdfs were disjoint. Which isn't the case.
We need a more general solution.

It seems that some error accumulates the higher the ray depth is.
For up until depth 7 the image produced will be as one would expect.
Clean, with very minimal noise, getting more nuanced by each depth increase.
However, there seems to be some bug with the PDF mixing. And I am sure
it is with the mixing, because either direct light sampling or the normal
sampling works without issue.
The mixed one though will produce this fairy fire type of effect, when
certain pixels would get completely blown out in one color. Not sure
what may be causing it. Gotta figure it out though.