"Fix" fairy fire
Went back to the original implementation and it behaves exactly
as the one in the rewrite. So the issue isn't with the implementaiton
but with the logic. This is not something that will be addressed at
this time.
However as a rule of thumb using a sampleCount at least 10 times that
of rayDepth will minimize the effect. This means that for a practical
final render, it won't pose a problem.
Eventually this should be fixed, though.