Add command line arguments handling
Huh, this was a pain. Wanted to use an external library like cxxopt,
but it wasn't available as a bazel module and I really didn't want to
figure out what arcane spells do I need to scribe upon the screen
to make it happen with bazel.
So using getopt and not supporting windows. Tough life.
Anyway, rather happy with the result. Everything I wanted to be configurable
is, and it looks usable.

Clang tidy doesn't like a number of things, but what can I do.
C++ shenanigans.