Add MeshCollision struct
This struct aims to hold additional information about the mesh collision
not just what is needed by the rest of the pipeline.
For example to enable the Sphere's for Gourand shading we need to know
which vertex was hit, so we can find it's neighbours and do the necessary
calculations.