Add Gouraud shading for the sphere mesh type
The implementation still has some bugs. Primarily that when
the original icosahedron mesh is subdivided, there seems to be some
bug in the division algorithm.
And some vertexes in the subdivided mesh won't have exactly 6 trinagles
being connected to. This will cause issues in the normal calculations
and is obviously a sign of a bug that needs to be fixed.