Refact: Extract mesh data structure
Vertices, triangles and all the data that is associated with these
is a mesh. Extracting this into it's separate struct makes the code more
readeable.

Additionally, forced all geometries to be only constructable through
their respective buildXYZ function. The reason is that, the cleanest
way constructing any of them is to build a default one at the origin,
with default sizes, than scale/translate them as necessary. This
translation then can update all the necessary support structures about
the mesh. Now logic is simple. There is only one flow, and there is no
need to think about any edge cases.

Furthermore with this change it is becoming more and more obvious that
a shape is nothing more than a mesh. We only need a shared interface
to override those few operations that may need it. Even this though
may be factored away in the future.