Remove Gouraud shading from sphere
Well, this sucks. I thought it would help, but it won't. Simply, just
tesselate the sphere more. Trying to mess with the normals is a bad
idea for ray tracing. It just has a lot of calculation issues
especially on the edge of the object.

Additionally renamed the sphere to icosphere, because it seems
this is the pattern, and it hopefully signifies to those more
knowledgeable than I, not to try and apply shading to it.

Now that I think about it, in ray tracing we probabbly shouldn't
ever mess with the normals.