Add bounding box to icosphere
This can be used to speed up collision detection in a substantial
way. Checking if we have hit a 6 sided cuboid is massively cheaper
then checking if we have hit any of the hundreds of triangles making
up an icosphere. Additionally this bounding cuboid will be transformed
along the initial icosphere, so it will maintain it's optimal size.

Took me a while to get to this point, because I didn't fully realize
the implications of what a good default size is for the objects is.
Particularly got confused by the myriad of issues I have made mixing
up diameters and radiuses, widths and half widths and especially
expectations on how scaling works. The most egregious error I have made
was not realizing that only something with a unit length will scale
in an expected way in any situation. For the bounding box to be able
to track the original sphere, it has to use unit distances from it's
center.