Change linear_partition rendering to pixel_partition
Originally the image was chunked into lines, but this acted very
inefitienlty. The threads that would receive the parts of the picture
that contained a lot of complex elements, like dielectric icospheres,
would work much longer than other threads. So often times, most threads
would be complete while 2 or 3 would still be working away, sometimes
almost as long as the rest had to.
This isn't ideal at all, the image is now separated between threads
on a pixel basis. This way the threads can share the load as ideally
as possible.

This can give speedups of 2-3x.