Insert 3D DDA implementation into the render pipeline
The current implementation doesn't handle properly if the ray starts
from the outside of the scene, so as a shortcut we just take an x
number of steps at a maximum.
Additionally, voxelization seems to be very slooow! Have to make it
multithreaded for sure. Can't even finish just the voxelization for
the floating-sphere scene in 5 mins on one thread.